feat/minimal-openapi-migration

+++ src/Movies.Contracts/Responses/Movies/MovieResponse.cs
@@ -16,7 +16,7 @@ public record MovieResponse {
   [SwaggerSchema("The title of movie", Nullable = false)]
   public string? Title { get; set; }
   [SwaggerSchema("The year of release", Nullable = false)]
-  public long? YearOfRelease { get; set; }
+  public short? YearOfRelease { get; set; }
   [SwaggerSchema("The slug of movie", Nullable = false)]
   public string? Slug { get; set; }
